requirements.txt
anthropic==0.45.0
pydantic==2.5.3
requests==2.31.0
python-dotenv==1.0.1
langchain
langgraph
langchain-anthropic